fix: remove or select base views

+++ shared/src/containers/Views/ViewsMenuContainer/BaseViewsTags.tsx
@@ -5,7 +5,8 @@ import { confirmDialog } from 'primereact/confirmdialog'
 import { usePowerpack } from '@shared/context'
 import { SectionHeader } from '@shared/containers/Views/ViewsMenu/SectionHeader'
 import { useLocalStorage } from '@shared/hooks'
-
+import { Icon } from '@ynput/ayon-react-components'
+import clsx from 'clsx'
 
 const BaseViewsTagContainer: FC = () => {
   const {
@@ -13,30 +14,38 @@ const BaseViewsTagContainer: FC = () => {
     studioBaseView,
     onCreateBaseView,
     onDeleteBaseView,
+    onUpdateWorkingView,
   } = useViewsContext()
 
   const { powerLicense, setPowerpackDialog } = usePowerpack()
   const [collapsed, setCollapsed] = useLocalStorage('collapsed-default-view', false)
 
-  const handleBaseViewAction = async (isStudioScope: boolean) => {
+  const handleBaseViewAction = async (isStudioScope: boolean, remove?: boolean) => {
     const existingBase = isStudioScope ? studioBaseView : projectBaseView
 
-    if(!isStudioScope && !powerLicense ){
+    if (!isStudioScope && !powerLicense) {
       setPowerpackDialog('sharedViews')
       return
     }
 
     if (existingBase) {
-      confirmDialog({
-        message: `Are you sure you want to remove this default view?`,
-        header: `Remove Default View`,
-        acceptLabel: 'Remove',
-        rejectLabel: 'Cancel',
-        accept: async () => {
-          await onDeleteBaseView(existingBase.id as string, isStudioScope)
-        }
-      })
+      if (remove) {
+        // remove button clicked, ask to remove the base view
+        confirmDialog({
+          message: `Are you sure you want to remove this default view?`,
+          header: `Remove Default View`,
+          acceptLabel: 'Remove',
+          rejectLabel: 'Cancel',
+          accept: async () => {
+            await onDeleteBaseView(existingBase.id as string, isStudioScope)
+          },
+        })
+      } else {
+        // set the working view to the existing base view
+        onUpdateWorkingView({ settings: existingBase.settings }, { selectView: true })
+      }
     } else {
+      // create new base view
       await onCreateBaseView(isStudioScope)
     }
   }
@@ -48,19 +57,19 @@ const BaseViewsTagContainer: FC = () => {
         collapsed={collapsed}
         id="default-views"
         title="Default views"
-        style={{marginBottom: '10px'}}
+        style={{ marginBottom: '10px' }}
       />
       {!collapsed && (
         <Styled.BaseViewsContainer>
           <ScopeIcon
             existingView={!!studioBaseView}
             label={'Studio'}
-            onClick={() => handleBaseViewAction(true)}
+            onClick={(r) => handleBaseViewAction(true, r)}
           />
           <ScopeIcon
             existingView={!!projectBaseView}
             label={'Project'}
-            onClick={() => handleBaseViewAction(false)}
+            onClick={(r) => handleBaseViewAction(false, r)}
             powerLicense={powerLicense}
           />
         </Styled.BaseViewsContainer>
@@ -72,7 +81,7 @@ export default BaseViewsTagContainer
 
 type ScopeIconProps = {
   existingView: boolean
-  onClick: () => void
+  onClick: (remove?: boolean) => void
   label: string
   powerLicense?: boolean
 }
@@ -88,14 +97,24 @@ const ScopeIcon: FC<ScopeIconProps> = ({
     onClick()
   }
 
-  const className = powerLicense === false ? 'powerpack-locked' : existingView ? 'active' : ''
+  const handleRemove = (e: React.MouseEvent) => {
+    e.stopPropagation()
+    onClick(true)
+  }
 
   return (
     <Styled.ViewButton
-      label={label}
-      className={className}
+      className={clsx({
+        ['powerpack-locked']: powerLicense === false,
+        ['active']: existingView,
+      })}
       onClick={handleClick}
-      icon={powerLicense === false ? "bolt" : existingView ? 'close' : 'add'}
-    />
+    >
+      <Icon
+        icon={powerLicense === false ? 'bolt' : existingView ? 'close' : 'add'}
+        onClick={(e) => existingView && handleRemove(e)}
+      />
+      {label}
+    </Styled.ViewButton>
   )
 }
